fix upsert
