+++ server/utils/db-helpers.js
@@ -104,10 +104,10 @@ const insertRecord = async (conn, tableName, values) => {
 
 const upsertRecord = async (conn, tableName, values, uniqueBy = 'id') => {
   // check for existance first
-  const existing = await conn.query(`SELECT * FROM ${tableName} WHERE ? = ?`, [
-    uniqueBy,
-    values[uniqueBy],
-  ]);
+  const existing = await conn.query(
+    `SELECT * FROM ${tableName} WHERE ${uniqueBy} = ?`,
+    values[uniqueBy]
+  );
   if (existing.length > 0) {
     return updateRecord(conn, tableName, existing[0].id, values);
   }
